Primary Commit 1. Authored By Krister Blanch.
This is a barebones build to provide a plan-position indicator (PPI) for a Navico Radar.
It requires X11.
OpenCluon and OpenDLV are header only, and included with this repo
The README is only partially complete. System requires an openDLV session broadcasting the DLV Radar message (CURRENTLY NOT A PART OF THE STANDARD DLV MESSAGE SET.)
If using this with other openDLV microservices, use the openDLV standard message set included with this software when compiling the other microservices.

Code comment is patchy and requires updating.