Optic Flow Commit 3. Authored by Krister Blanch.
<Fixes>
No Fixes.

<Features>
Introduced libcester as a testing framework. Not entirely sold on it, as it is struggling to merge with cmake.
Running the cester test results in a seperation of the tests with correct values, but running it as part of the CI/CD or Cmake makes it treat the whole exe as
a single test.

<New/Ongoing Bugs>
Same as prior commit.

<Compilability/Status>
Compiles locally. Still having some issues with Docker. Considering moving the entire CI/CD to Docker to focus on a single deployment.