Optic Flow Commit 9. Authored by Krister Blanch.
<Fixes>
Updated readme to include methods for building in docker and from source, and how to run a demonstration of the software.

<Features>
created a docker-compose file, and started working on a dockerfile for an Ubuntu build. Whole system can now be deployed via a single terminal command (docker-compose up)

<New/Ongoing Bugs>
Alpine Docker Build compiles but Cluon cannot communicate outside the container. Fine for unit testing, but the system cannot be used in deployment.

<Compilability/Status>
Compiles with Docker and from source. Can deploy in a demonstrator and a working method from source. Passes unit test in Docker and in source.