Optic Flow Commit 10. Authored by Krister Blanch.
<Fixes>
Cleaned up the formatting and typeface of the readme.md.

<Features>
Now with even more testing! Completely fleshed out the testing suite with a dozen test cases. Succesfully tests in both docker and local.

<New/Ongoing Bugs>

<Compilability/Status>
Compiles with Docker and from source. Can deploy in a demonstrator and a working method from source. Passes unit test in Docker and in source.
-wall is throwing a lot of warnings.